setup project file

+++ src/utils/cloudinary.js
@@ -16,11 +16,16 @@ import fs from "fs";
         const response = await cloudinary.uploader.upload(localFilePath, {
             resource_type: "auto", // Automatically determine the resource type (image, video, etc.)
         });
-        console.log("Cloudinary upload response:", response.url);
+        // console.log("Cloudinary upload response:", response.url);
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        } // Clean up the local file after upload
         return response;
     } catch (error) {
-        fs.unlinkSync(localFilePath); // Clean up the local file if upload fails
+       if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        } // Clean up the local file if upload fails
         return null; // Return null if upload fails
     }
 };
-export default uploadOnCloudinary;
+export {uploadOnCloudinary};
